Move custom hook useGetPoll to correct directory
